migrate simple pages
