+++ app/sessions/page.page.tsx
@@ -1,3 +1,4 @@
+'uce client'
 import dayjs from 'dayjs'
 import { useRouter } from 'next/router'
 import { useEffect } from 'react'
