+++ app/settings/page.page.tsx
@@ -1,0 +1,12 @@
+import Grid from '@mui/material/Grid'
+import SessionRedirectSwitch from '../../components/SessionRedirectSwitch'
+
+export default function SettingsPage() {
+  return (
+    <Grid container spacing={2}>
+      <Grid alignItems="center" display="flex" size={12}>
+        <SessionRedirectSwitch />
+      </Grid>
+    </Grid>
+  )
+}
